+++ src/lib/FormGroupLayout.tsx
@@ -1,22 +1,26 @@
-import { PropsWithChildren, ReactNode, CSSProperties } from "react";
+/* eslint-disable complexity */
+import { PropsWithChildren, ReactNode, CSSProperties, useMemo } from "react";
 import { FieldError, FieldValues, get } from "react-hook-form";
 import { FormGroup, FormFeedback, FormText, InputGroup } from "reactstrap";
 import { useSafeNameId } from "src/lib/hooks/useSafeNameId";
-import { CommonInputProps } from "./types/CommonInputProps";
+import { CommonInputProps, MergedAddonProps } from "./types/CommonInputProps";
 import "./styles/FormGroupLayout.css";
 import { FormGroupLayoutLabel } from "./FormGroupLayoutLabel";
 import { useFormContext } from "./context/FormContext";
 
-interface FormGroupLayoutProps<T extends FieldValues>
+interface FormGroupLayoutProps<T extends FieldValues, TRenderAddon>
   extends PropsWithChildren<Pick<CommonInputProps<T>, "helpText" | "label" | "name" | "id" | "labelToolTip" | "inputOnly">> {
   layout?: "checkbox" | "switch";
-  addonLeft?: ReactNode;
-  addonRight?: ReactNode;
+  addonLeft?: ReactNode | ((props: TRenderAddon) => ReactNode);
+  addonRight?: ReactNode | ((props: TRenderAddon) => ReactNode);
+  addonProps?: MergedAddonProps<TRenderAddon>;
   inputGroupStyle?: CSSProperties;
+  formGroupId?: string;
 }
 
-const FormGroupLayout = <T extends FieldValues>(props: FormGroupLayoutProps<T>) => {
-  const { label, helpText, children, layout, labelToolTip, inputOnly, addonLeft, addonRight, inputGroupStyle } = props;
+const FormGroupLayout = <T extends FieldValues, TRenderAddon = unknown>(props: FormGroupLayoutProps<T, TRenderAddon>) => {
+  const { label, helpText, children, layout, labelToolTip, inputOnly, addonLeft, addonRight, inputGroupStyle, formGroupId, addonProps } =
+    props;
   const { name, id } = useSafeNameId(props.name, props.id);
   const {
     formState: { errors },
@@ -32,13 +36,29 @@ const FormGroupLayout = <T extends FieldValues>(props: FormGroupLayoutProps<T>) 
     throw "'inputOnly' is not possible with switches or checkboxes";
   }
 
+  const effectiveAddonLeft = useMemo(
+    () =>
+      addonLeft instanceof Function && addonProps
+        ? (addonLeft as (props: TRenderAddon) => ReactNode)(addonProps)
+        : (addonLeft as ReactNode),
+    [addonLeft, addonProps],
+  );
+
+  const effectiveAddonRight = useMemo(
+    () =>
+      addonRight instanceof Function && addonProps
+        ? (addonRight as (props: TRenderAddon) => ReactNode)(addonProps)
+        : (addonRight as ReactNode),
+    [addonRight, addonProps],
+  );
+
   return inputOnly ? (
     <>
       {children}
       <FormFeedback>{errorMessage}</FormFeedback>
     </>
   ) : (
-    <FormGroup switch={switchLayout ? true : undefined} check={checkboxLayout ? true : undefined}>
+    <FormGroup id={formGroupId} switch={switchLayout ? true : undefined} check={checkboxLayout ? true : undefined}>
       <FormGroupLayoutLabel<T> label={label} fieldName={name} fieldId={id} tooltip={labelToolTip} layout={layout} />
       {switchLayout || checkboxLayout ? (
         children
@@ -51,9 +71,9 @@ const FormGroupLayout = <T extends FieldValues>(props: FormGroupLayoutProps<T>) 
           }}
           className={fieldError ? "is-invalid" : undefined}
         >
-          {addonLeft}
+          {effectiveAddonLeft}
           {children}
-          {addonRight}
+          {effectiveAddonRight}
         </InputGroup>
       )}
       <FormFeedback>{errorMessage}</FormFeedback>